refactor: rename media selection components
